Add prescaler for ADC

+++ Code/line-follower/uart.h
@@ -6,4 +6,4 @@ void UART_init();
 
 // UART andmed massiivides, ilma katkestuseta
 void UART_write(char *data);
-void motor_drive(uint8_t left, uint8_t right);
+void motor_drive(int16_t left, int16_t right);
